I hope this is fixed... pls read description
Things I've done:
Re-added majority of the prefabs with the correct scripts.
Added in the "new turrets" that are in the folder called "final turret". They still need to be fixed themselves as their rotation point is incorrect.
I've also done some house cleaning with old prefabs, if I have deleted any LET ME KNOW and I'll get around to it. 
If this commit is still broken then I'll be having a sleepless night...